[fix] use parametric-types to speed up compilation speed of pf
Before this, on SBCL 2.2.4 and later, compilation times of pf become
impractical. Even if one were to blame SBCL for recent changes, there
exist other instances wherein even older SBCL versions take too long for
compiling large functions. In essence, it is not a good idea to rely on a
compiler being able to compile large functions for things to keep working
correctly. Instead, we adopt this approach, wherein we reduce the number of
polymorphs from 66 to 11; this also resulted in a significant increase in the
runtime performances where relevant.